# backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: YolNext API
  description: Comprehensive logistics and transportation management platform API
  version: 1.0.0
  contact:
    name: YolNext Support
    email: support@yolnext.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.yolnext.com
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /api/shipments:
    get:
      tags:
        - Shipments
      summary: Get shipments
      description: Retrieve shipments based on user role and filters
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_progress, delivered, cancelled]
        - name: search
          in: query
          schema:
            type: string
            description: Search by tracking number, pickup city, or delivery city
      responses:
        '200':
          description: Shipments retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Shipment'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Shipments
      summary: Create new shipment
      description: Create a new shipment request
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateShipmentRequest'
      responses:
        '201':
          description: Shipment created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Shipment created successfully
                  data:
                    $ref: '#/components/schemas/Shipment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: user_123
        name:
          type: string
          example: John Doe
        email:
          type: string
          format: email
          example: john@example.com
        phone:
          type: string
          example: "+905551234567"
        userType:
          type: string
          enum: [individual, corporate, nakliyeci, tasiyici, admin]
          example: individual
        company:
          type: string
          example: ABC Transport
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: 2024-01-01T10:00:00Z
        updatedAt:
          type: string
          format: date-time
          example: 2024-01-01T10:00:00Z
    Shipment:
      type: object
      properties:
        id:
          type: string
          example: shipment_123
        trackingNumber:
          type: string
          example: TRK001234567
        from:
          type: object
          properties:
            address:
              type: string
              example: İstanbul, Turkey
            city:
              type: string
              example: İstanbul
            coordinates:
              type: object
              properties:
                lat:
                  type: number
                  example: 41.0082
                lng:
                  type: number
                  example: 28.9784
        to:
          type: object
          properties:
            address:
              type: string
              example: Ankara, Turkey
            city:
              type: string
              example: Ankara
            coordinates:
              type: object
              properties:
                lat:
                  type: number
                  example: 39.9334
                lng:
                  type: number
                  example: 32.8597
        status:
          type: string
          enum: [pending, in_progress, delivered, cancelled]
          example: pending
        priority:
          type: string
          enum: [low, normal, high]
          example: normal
        weight:
          type: number
          example: 100.5
        volume:
          type: number
          example: 2.5
        value:
          type: number
          example: 1500.00
        pickupDate:
          type: string
          format: date-time
          example: 2024-01-15T10:00:00Z
        deliveryDate:
          type: string
          format: date-time
          example: 2024-01-16T18:00:00Z
        notes:
          type: string
          example: Fragile items
        driver:
          $ref: '#/components/schemas/Driver'
        shipper:
          $ref: '#/components/schemas/User'
        createdAt:
          type: string
          format: date-time
          example: 2024-01-01T10:00:00Z
        updatedAt:
          type: string
          format: date-time
          example: 2024-01-01T10:00:00Z
    CreateShipmentRequest:
      type: object
      required:
        - from
        - to
        - weight
        - pickupDate
        - deliveryDate
      properties:
        from:
          type: object
          required:
            - address
            - city
          properties:
            address:
              type: string
              example: İstanbul, Turkey
            city:
              type: string
              example: İstanbul
            coordinates:
              type: object
              properties:
                lat:
                  type: number
                lng:
                  type: number
        to:
          type: object
          required:
            - address
            - city
          properties:
            address:
              type: string
              example: Ankara, Turkey
            city:
              type: string
              example: Ankara
            coordinates:
              type: object
              properties:
                lat:
                  type: number
                lng:
                  type: number
        weight:
          type: number
          minimum: 0
          example: 100.5
        volume:
          type: number
          minimum: 0
          example: 2.5
        value:
          type: number
          minimum: 0
          example: 1500.00
        pickupDate:
          type: string
          format: date-time
          example: 2024-01-15T10:00:00Z
        deliveryDate:
          type: string
          format: date-time
          example: 2024-01-16T18:00:00Z
        priority:
          type: string
          enum: [low, normal, high]
          default: normal
        notes:
          type: string
          example: Fragile items
    Driver:
      type: object
      properties:
        id:
          type: string
          example: driver_123
        name:
          type: string
          example: Ahmet Yılmaz
        phone:
          type: string
          example: "+905551234567"
        vehicle:
          type: object
          properties:
            type:
              type: string
              example: Truck
            plate:
              type: string
              example: 34ABC123
            capacity:
              type: number
              example: 5000
        status:
          type: string
          enum: [available, busy, offline]
          example: available
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
        pages:
          type: integer
          example: 5
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Error message
        errors:
          type: array
          items:
            type: string
          example: [field1 error, field2 error]
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
            example:
              success: false
              message: Authentication required
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
            example:
              success: false
              message: Validation failed
              errors: [email is required, password must be at least 6 characters]
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token
...